# academify/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Academify API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Student Community & Collaboration Platform — COMP6703001"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://e2526-wads-b4ac-02.csbihub.id",
      "description": "Production server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "HTTP-only session cookie set by Better Auth after sign-in. Sign in via the app or POST /api/auth/sign-in/email, then use Try it out from the same origin."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Legacy alias — Academify uses session cookies, not Bearer JWT."
      }
    },
...
  },
  "paths": {
...
    "/api/reports": {
      "get": {
        "summary": "Get all reports (Moderator/Admin only)",
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "sessionCookieAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "reviewed",
                "resolved",
                "dismissed"
              ]
            },
            "description": "Filter by report status"
          }
        ],
        "responses": {
          "200": {
            "description": "List of reports"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden - Moderator/Admin only"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "summary": "Submit a new report",
        "tags": [
          "Reports"
        ],
        "security": [
          {
            "sessionCookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "targetType",
                  "targetId",
                  "reason"
                ],
                "properties": {
                  "targetType": {
                    "type": "string",
                    "enum": [
                      "post",
                      "comment",
                      "user"
                    ],
                    "example": "post"
                  },
                  "targetId": {
                    "type": "string",
                    "example": "1"
                  },
                  "reason": {
                    "type": "string",
                    "example": "This post contains spam."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Report submitted successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "401": {
            "description": "Not authenticated"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}